moved spec files to tests directory
