+++ tests/client/user-home.spec.js
